backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ultralytics==8.0.196
pillow==10.0.1
python-multipart==0.0.6
